backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
pydantic
sqlmodel
langchain-chroma
requests
